Fix seeking in dumchar driver.
The dumchar driver offers access to several areas of the main disk as
character devices. To do this, in some cases its operations delegate on
the operations of the main disk. However, when seeking, this is not
possible; an explicit implementation is needed instead. Otherwise, the
seek would be based on the size of the main disk instead of the size of
the area exposed by the character device, which would cause the seek
operation to behave incorrectly.

This change implicitly fixes the use of the "fseek" function from user
space with devices managed by the dumchar driver, as it could not be
used to get the current position in the device file, set it to a
different position nor get the size of the device file.