Fix updating the internal file position in dumchar driver.
The dumchar driver offers access to several areas of the main disk as
character devices. To do this, it wraps an internal file pointer in the
private data of the exposed file pointer. The position of the internal
file pointer was modified at the end of the read and write operations on
the exposed file pointer because, typically, the position of the exposed
file pointer is modified after the read and write operations. However,
it is not always the case; in some cases after the read and write
operations the position of the exposed file pointer is left untouched.
As the position of the internal file pointer is used in some operations
this resulted in an incorrect behavior. Therefore, now the position of
the internal file pointer is updated only before being used based on the
position of the exposed file pointer, as it is the one updated by the
callers of its operations and thus holds the correct value.

This change implicitly fixes the use of the "pread" and "pwrite"
functions from user space with devices managed by the dumchar driver, as
the previous implementation caused the file position to be modified when
those functions were used.